# backend/cmd/order-service/docs/swagger.yaml
basePath: /
definitions:
  domain.Order:
    properties:
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      items:
        items:
          $ref: '#/definitions/domain.OrderItem'
        type: array
      shipping_address:
        $ref: '#/definitions/domain.ShippingAddress'
      status:
        $ref: '#/definitions/domain.OrderStatus'
      total_amount:
        description: cents
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  domain.OrderItem:
    properties:
      id:
        type: string
      order_id:
        type: string
      price:
        description: unit price in cents
        type: integer
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  domain.OrderStatus:
    enum:
    - pending
    - confirmed
    - paid
    - cancelled
    - refunded
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusConfirmed
    - StatusPaid
    - StatusCancelled
    - StatusRefunded
  domain.PaymentStatus:
    enum:
    - pending
    - processing
    - succeeded
    - failed
    - refunded
    type: string
    x-enum-varnames:
    - PaymentPending
    - PaymentProcessing
    - PaymentSucceeded
    - PaymentFailed
    - PaymentRefunded
...
  domain.ShippingAddress:
    properties:
      city:
        type: string
      country:
        description: ISO 3166-1 alpha-2, e.g. "US"
        type: string
      name:
        type: string
      postal_code:
        type: string
      state:
        type: string
      street:
        type: string
    type: object
...
  handler.CancelOrderResponse:
    properties:
      message:
        example: order cancelled
        type: string
      order_id:
        example: ord-abc-123
        type: string
    type: object
...
  handler.CreateOrderBody:
    properties:
      currency:
        example: USD
        type: string
      items:
        items:
          $ref: '#/definitions/handler.CreateOrderItemBody'
        minItems: 1
        type: array
      shipping_address:
        $ref: '#/definitions/handler.ShippingAddressBody'
    required:
    - currency
    - items
    type: object
  handler.CreateOrderItemBody:
    properties:
      product_id:
        example: prod-abc-123
        type: string
      quantity:
        example: 2
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: something went wrong
        type: string
    type: object
...
  handler.ListOrdersResponse:
    properties:
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      orders:
        items:
          $ref: '#/definitions/domain.Order'
        type: array
    type: object
...
  handler.PaymentErrorResponse:
    properties:
      error:
        example: payment not found
        type: string
    type: object
  handler.PaymentResponse:
    properties:
      amount:
        example: 5000
        type: integer
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      currency:
        example: USD
        type: string
      failure_reason:
        example: card_declined
        type: string
      id:
        example: pay-uuid
        type: string
      method:
        example: card
        type: string
      order_id:
        example: order-uuid
        type: string
      status:
        allOf:
        - $ref: '#/definitions/domain.PaymentStatus'
        example: succeeded
      transaction_id:
        example: txn_abc123
        type: string
      updated_at:
        example: "2026-01-01T00:01:00Z"
        type: string
      user_id:
        example: user-uuid
        type: string
    type: object
...
  handler.RefundResponse:
    properties:
      message:
        example: payment refunded
        type: string
      order_id:
        example: ord-uuid
        type: string
      payment_id:
        example: pay-uuid
        type: string
    type: object
...
  handler.ShippingAddressBody:
    properties:
      city:
        example: New York
        type: string
      country:
        example: US
        type: string
      name:
        example: Alice Smith
        type: string
      postal_code:
        example: "10001"
        type: string
      state:
        example: NY
        type: string
      street:
        example: 123 Main St
        type: string
    required:
    - city
    - country
    - name
    - postal_code
    - street
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Manages order creation, retrieval, and status updates driven by payment
    events.
  title: Order Service
  version: "1.0"
paths:
  /api/v1/orders:
    get:
      description: Returns paginated orders for the authenticated user.
      parameters:
      - default: 20
        description: Max results (default 20, max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListOrdersResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: 'Create a new order. Idempotent: same Idempotency-Key returns the
        existing order.'
      parameters:
      - description: Client-generated unique key for idempotency
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Order details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOrderBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create order
      tags:
      - orders
  /api/v1/orders/{id}:
    delete:
      description: Cancels a pending or confirmed order. Only the owner can cancel.
        Paid or refunded orders cannot be cancelled.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CancelOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel order
      tags:
      - orders
    get:
      description: Returns an order. Only the owner can access it.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Order'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - orders
  /api/v1/payments/{id}:
    get:
      description: Returns payment details including status, amount, and transaction
        ID.
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.PaymentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payment
      tags:
      - payments
  /api/v1/payments/{id}/refund:
    post:
      description: Refunds a succeeded payment. Only the payment owner can request
        a refund. Publishes payments.refunded which credits the user's wallet.
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RefundResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
      security:
      - BearerAuth: []
      summary: Refund payment
      tags:
      - payments
  /api/v1/payments/order/{order_id}:
    get:
      description: Returns the payment associated with a specific order ID.
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.PaymentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.PaymentErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payment by order
      tags:
      - payments
...
securityDefinitions:
  BearerAuth:
    description: JWT token issued by the API Gateway
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"